+++ src/components/Common/styles.js
@@ -1,0 +1,6 @@
+import styled from "styled-components";
+
+export const Container = styled.div`
+  height: 100vh;
+  width: auto;
+`;
