+++ src/components/Content/Content.js
@@ -1,8 +1,8 @@
 import React from "react";
 import { ContentContainer } from "../../pages/styles";
 
-const Body = ({ children }) => {
+const Content = ({ children }) => {
   return <ContentContainer>{children}</ContentContainer>;
 };
 
-export default Body;
+export default Content;
