+++ src/components/Content/index.js
@@ -1,1 +1,1 @@
-export * from "./Content";
+export { default as Content } from "./Content";
