+++ src/components/Header/index.js
@@ -1,1 +1,3 @@
-export * from "./Header";
+export { default as Header } from "./Header";
+export { default as HamburgerMenu } from "./HamburgerMenu";
+export { default as Search } from "./Search";
