+++ src/components/Layout/index.js
@@ -1,0 +1,2 @@
+export { default as TaskLayout } from "./TaskLayout";
+export { default as SettingLayout } from "./SettingLayout";
