--- src/components/SettingLayout.js/index.js
@@ -1,1 +1,0 @@
-export * from "./SettingLayout";
