+++ src/components/SideBar/index.js
@@ -1,1 +1,1 @@
-export * from "./SideBar";
+export { default as SideBar } from "./SideBar";
