+++ src/components/TaskContent/index.js
@@ -1,1 +1,2 @@
 export { default as TaskContent } from "./TaskContent";
+export { default as TaskSiderBar } from "./TaskSiderBar";
