--- src/components/TaskLayout/index.js
@@ -1,1 +1,0 @@
-export { default as TaskLayout } from "./TaskLayout";
