+++ src/components/index.js
@@ -1,8 +1,3 @@
 export { default as Search } from "./Header/Search";
 
 export { default as SideBar } from "./SideBar/SideBar";
-export { default as Task } from "./Task/Task";
-
-export { default as Title } from "./Header/Title";
-export { default as Body } from "./Body/Body";
-export { default as Header } from "./Header/Header";
